+++ CHANGES.md
@@ -3,7 +3,9 @@ Version 0.2.4 released 2012-7-??
 
 * Added unit-tests for SPMDVariation
 * Added 'mpirun' support for local job plug-in (via SPMDVariation)
-* Added missing flags to filesystem package
+* Added some of the missing methods and flags to filesystem package
+* An URL object can now be constructed from another URL object
+
 
 Version 0.2.3 released 2012-6-26
 ----------------------------------------------------------------------